refactor (registration-api-client) Move ApiRegistrationPublic class definition to no.fdk.registration.common
