+++ applications/informationmodel-cat/pom.xml
@@ -95,7 +95,7 @@
         <dependency>
             <groupId>no.dcat</groupId>
             <artifactId>registration-api-client</artifactId>
-            <version>1.5.0</version>
+            <version>1.6.0</version>
         </dependency>
         <dependency>
             <groupId>no.dcat</groupId>
