+++ applications/registration-api/pom.xml
@@ -202,8 +202,8 @@
         </dependency>
         <dependency>
             <groupId>no.dcat</groupId>
-            <artifactId>registration-api-client</artifactId>
-            <version>1.5.0</version>
+            <artifactId>registration-api-common</artifactId>
+            <version>1.0.0</version>
         </dependency>
 
         <dependency>
