+++ libraries/registration-api-client/src/main/java/no/dcat/client/registrationapi/RegistrationApiClient.java
@@ -6,6 +6,7 @@ import com.google.gson.JsonParser;
 import com.google.gson.reflect.TypeToken;
 import lombok.Getter;
 import lombok.Setter;
+import no.fdk.registration.common.ApiRegistrationPublic;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.web.client.RestTemplate;
